Checked and fixed some bugs, edited makefile to work with windows, added dataset file for dns

+++ Makefile
@@ -9,10 +9,19 @@ all :
 	@$(MAKE) -s detector
 
 dataset : clean
+ifeq ($(OS),Windows_NT)
+	mkdir dataset
+else
+	mkdir -p dataset
+endif
 	python3 makedataset.py
 detector :
 	python main.py --dataset dataset
 single :
 	python main.py --input_pcap dataset/out-19.pcap -l debug
 clean :
+ifeq ($(OS),Windows_NT)
+	@if exist dataset rmdir /s /q dataset
+else
 	rm -rf dataset
+endif
